handle drag & drop in Main.vue
