+++ app/components/windows/Main.vue
@@ -1,5 +1,5 @@
 <template>
-<div class="main" :class="{'night-theme': nightTheme, 'day-theme': !nightTheme}" id="mainWrapper">
+<div class="main" :class="{'night-theme': nightTheme, 'day-theme': !nightTheme}" id="mainWrapper" @drop="onDropHandler">
   <title-bar :title="title" />
   <div class="main-spacer bgColor-teal"></div>
   <div class="main-contents">
